+++ lib/src/v3/api/websockets.dart
@@ -569,8 +569,8 @@ class WsEventGetPrivateMessages extends WsEvent<List<PrivateMessageView>> {
   const WsEventGetPrivateMessages(List<PrivateMessageView> data) : super(data);
 }
 
-class WsEventGetReportCount extends WsEvent<GetReportCountResponse> {
-  const WsEventGetReportCount(GetReportCountResponse data) : super(data);
+class WsEventGetReportCount extends WsEvent<ReportCount> {
+  const WsEventGetReportCount(ReportCount data) : super(data);
 }
 
 class WsEventResolveObject extends WsEvent<ResolveObjectResponse> {
@@ -589,6 +589,6 @@ class WsEventBlockCommunity extends WsEvent<BlockedCommunity> {
   const WsEventBlockCommunity(BlockedCommunity data) : super(data);
 }
 
-class WsEventGetUnreadCount extends WsEvent<GetUnreadCountResponse> {
-  const WsEventGetUnreadCount(GetUnreadCountResponse data) : super(data);
+class WsEventGetUnreadCount extends WsEvent<UnreadCount> {
+  const WsEventGetUnreadCount(UnreadCount data) : super(data);
 }