+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Added
 
 - Add reporting API
-- Add `GetUnreadCounts` endpoint
+- Add `GetUnreadCount` endpoint
 - Add `honeypot` fields
 
 ### Changed